update optical frame publisher args
