+++ submitted_models/robotika_x2_sensor_config_1/launch/vehicle_topics.launch
@@ -81,14 +81,11 @@
     <node
       pkg="subt_ros"
       type="optical_frame_publisher"
-      name="optical_frame_publisher_camera_info"
-      args="camera_info sensor_msgs/CameraInfo">
-    </node>
-    <node
-      pkg="subt_ros"
-      type="optical_frame_publisher"
-      name="optical_frame_publisher_camera"
-      args="image_raw sensor_msgs/Image">
+      name="optical_frame_publisher">
+      <remap from="input/image" to="image_raw" />
+      <remap from="output/image" to="optical/image_raw" />
+      <remap from="input/camera_info" to="camera_info" />
+      <remap from="output/camera_info" to="optical/camera_info" />
     </node>
 
     <node
@@ -109,8 +106,10 @@
     <node
       pkg="subt_ros"
       type="optical_frame_publisher"
-      name="optical_frame_publisher_rgbd_depth"
-      args="front/depth sensor_msgs/Image">
+      name="optical_frame_publisher_depth"
+      args="depth">
+      <remap from="input/image" to="front/depth" />
+      <remap from="output/image" to="front/optical/depth" />
     </node>
 
     <!-- Gas sensor -->
