+++ submitted_models/ssci_x2_sensor_config_1/launch/vehicle_topics.launch
@@ -74,8 +74,10 @@
       <node
         pkg="subt_ros"
         type="optical_frame_publisher"
-        name="optical_frame_publisher_rgbd_depth"
-        args="front/depth sensor_msgs/Image">
+        name="optical_frame_publisher_depth"
+        args="depth">
+        <remap from="input/image" to="front/depth" />
+        <remap from="output/image" to="front/optical/depth" />
       </node>
 
       <node
@@ -96,14 +98,11 @@
       <node
         pkg="subt_ros"
         type="optical_frame_publisher"
-        name="optical_frame_publisher_camera_info"
-        args="front/camera_info sensor_msgs/CameraInfo">
-      </node>
-      <node
-        pkg="subt_ros"
-        type="optical_frame_publisher"
-        name="optical_frame_publisher_camera"
-        args="front/image_raw sensor_msgs/Image">
+        name="optical_frame_publisher">
+        <remap from="input/image" to="front/image_raw" />
+        <remap from="output/image" to="front/optical/image_raw" />
+        <remap from="input/camera_info" to="front/camera_info" />
+        <remap from="output/camera_info" to="front/optical/camera_info" />
       </node>
 
     </group>
@@ -127,8 +126,10 @@
       <node
         pkg="subt_ros"
         type="optical_frame_publisher"
-        name="optical_frame_publisher_rgbd_depth_up"
-        args="up/depth sensor_msgs/Image">
+        name="optical_frame_publisher_depth_up"
+        args="depth">
+        <remap from="input/image" to="up/depth" />
+        <remap from="output/image" to="up/optical/depth" />
       </node>
 
       <node
@@ -149,14 +150,11 @@
       <node
         pkg="subt_ros"
         type="optical_frame_publisher"
-        name="optical_frame_publisher_up_camera_info"
-        args="up/camera_info sensor_msgs/CameraInfo">
-      </node>
-      <node
-        pkg="subt_ros"
-        type="optical_frame_publisher"
-        name="optical_frame_publisher_up_camera"
-        args="up/image_raw sensor_msgs/Image">
+        name="optical_frame_publisher_up">
+        <remap from="input/image" to="up/image_raw" />
+        <remap from="output/image" to="up/optical/image_raw" />
+        <remap from="input/camera_info" to="up/camera_info" />
+        <remap from="output/camera_info" to="up/optical/camera_info" />
       </node>
 
     </group>
@@ -180,8 +178,10 @@
       <node
         pkg="subt_ros"
         type="optical_frame_publisher"
-        name="optical_frame_publisher_rgbd_depth_down"
-        args="down/depth sensor_msgs/Image">
+        name="optical_frame_publisher_depth_down"
+        args="depth">
+        <remap from="input/image" to="down/depth" />
+        <remap from="output/image" to="down/optical/depth" />
       </node>
 
       <node
@@ -202,14 +202,11 @@
       <node
         pkg="subt_ros"
         type="optical_frame_publisher"
-        name="optical_frame_publisher_down_camera_info"
-        args="down/camera_info sensor_msgs/CameraInfo">
-      </node>
-      <node
-        pkg="subt_ros"
-        type="optical_frame_publisher"
-        name="optical_frame_publisher_down_camera"
-        args="down/image_raw sensor_msgs/Image">
+        name="optical_frame_publisher_down">
+        <remap from="input/image" to="down/image_raw" />
+        <remap from="output/image" to="down/optical/image_raw" />
+        <remap from="input/camera_info" to="down/camera_info" />
+        <remap from="output/camera_info" to="down/optical/camera_info" />
       </node>
 
     </group>
@@ -249,28 +246,22 @@
       <node
         pkg="subt_ros"
         type="optical_frame_publisher"
-        name="optical_frame_publisher_stereo_camera_info_right"
-        args="front/right/camera_info sensor_msgs/CameraInfo">
-      </node>
-      <node
-        pkg="subt_ros"
-        type="optical_frame_publisher"
-        name="optical_frame_publisher_stereo_camera_info_left"
-        args="front/left/camera_info sensor_msgs/CameraInfo">
-      </node>
-      <node
-        pkg="subt_ros"
-        type="optical_frame_publisher"
-        name="optical_frame_publisher_stereo_camera_right"
-        args="front/right/image_raw sensor_msgs/Image">
-      </node>
-      <node
-        pkg="subt_ros"
-        type="optical_frame_publisher"
-        name="optical_frame_publisher_stereo_camera_left"
-        args="front/left/image_raw sensor_msgs/Image">
+        name="optical_frame_publisher_stereo_right">
+        <remap from="input/image" to="front/right/image_raw" />
+        <remap from="output/image" to="front/right/optical/image_raw" />
+        <remap from="input/camera_info" to="front/right/camera_info" />
+        <remap from="output/camera_info" to="front/right/optical/camera_info" />
       </node>
 
+      <node
+        pkg="subt_ros"
+        type="optical_frame_publisher"
+        name="optical_frame_publisher_stereo_left">
+        <remap from="input/image" to="front/left/image_raw" />
+        <remap from="output/image" to="front/left/optical/image_raw" />
+        <remap from="input/camera_info" to="front/left/camera_info" />
+        <remap from="output/camera_info" to="front/left/optical/camera_info" />
+      </node>
     </group>
 
     <!-- laser scan -->
